# specs/003-backend-http-db/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Backend API
  description: REST API for managing user tasks with JWT authentication
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.yourdomain.com
    description: Production server

paths:
  /api/v1/{user_id}/tasks:
    get:
      summary: List tasks for authenticated user
      description: Retrieve a list of tasks for the authenticated user with optional filtering
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user whose tasks to retrieve
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: Filter tasks by completion status (all, pending, completed)
          schema:
            type: string
            enum: [all, pending, completed]
            default: all
      responses:
        '200':
          description: Successfully retrieved tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - user_id in URL doesn't match JWT user
      security:
        - bearerAuth: []

    post:
      summary: Create a new task for authenticated user
      description: Create a new task associated with the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user creating the task
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - user_id in URL doesn't match JWT user
      security:
        - bearerAuth: []

  /api/v1/{user_id}/tasks/{task_id}:
    get:
      summary: Retrieve a specific task
      description: Get details of a specific task belonging to the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user whose task to retrieve
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          description: The ID of the task to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - user_id in URL doesn't match JWT user or task doesn't belong to user
        '404':
          description: Task not found
      security:
        - bearerAuth: []

    put:
      summary: Update a task
      description: Update the title and description of a specific task belonging to the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user whose task to update
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          description: The ID of the task to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - user_id in URL doesn't match JWT user or task doesn't belong to user
        '404':
          description: Task not found
      security:
        - bearerAuth: []

    delete:
      summary: Delete a task
      description: Delete a specific task belonging to the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user whose task to delete
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          description: The ID of the task to delete
          schema:
            type: string
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - user_id in URL doesn't match JWT user or task doesn't belong to user
        '404':
          description: Task not found
      security:
        - bearerAuth: []

  /api/v1/{user_id}/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion status
      description: Mark a task as complete or incomplete for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user whose task to update
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          description: The ID of the task to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCompleteRequest'
      responses:
        '200':
          description: Task completion status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid completion status
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - user_id in URL doesn't match JWT user or task doesn't belong to user
        '404':
          description: Task not found
      security:
        - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
        user_id:
          type: string
          description: ID of the user who owns this task
        title:
          type: string
          maxLength: 200
          description: Title of the task
        description:
          type: string
          nullable: true
          description: Optional description of the task
        completed:
          type: boolean
          description: Whether the task is completed
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated

    TaskCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 200
          description: Title of the task (max 200 characters)
        description:
          type: string
          nullable: true
          description: Optional description of the task

    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          description: New title for the task (max 200 characters)
        description:
          type: string
          nullable: true
          description: New description for the task

    TaskCompleteRequest:
      type: object
      required:
        - completed
      properties:
        completed:
          type: boolean
          description: Whether the task should be marked as completed